Extract map maker panel construction to factory classes
